Improve the primary monitor selection algorithm
Improve the primary monitor selection algorithm to look after the primary mark just on the connected monitors. If none of them has it, the first of them is selected as the primary. This behavior may technically not be correct, but it's unclear how likely is the target audience to use multimonitor setups. More feedback is necessary to determine the correct approach.